Shadow dancer. Perhaps overcomplicated.
